Lesson 8 and 9
LESSON 8 -SQL DataBase Design
PUT DATABASE filedata in one disk and another log in another disk
Ver Stored Procedures
Lesson 9- Prize Form Wire up
Connect code to sql
Function to connect to database

+++ TrackerLibrary/MatchupEntryModel.cs
@@ -16,7 +16,7 @@ namespace TrackerLibrary
         /// <summary>
         /// Represents the score for this particular team.
         /// </summary>
-        public double Score { get; set; }
+        public int Score { get; set; }
         /// <summary>
         /// Represents the matchup that this team came from as the winner.
         /// </summary>
